requirements.txt
ldap3
dnspython
pysmb
pycrypto
pysocks